Refactor Firestore class to use constructor params.
