+++ src/components/Firebase/firebase.js
@@ -1,23 +1,8 @@
-import app from "firebase/app";
-import "firebase/auth";
-import "firebase/firestore";
-
-const config = {
-  apiKey: process.env.REACT_APP_API_KEY,
-  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
-  databaseURL: process.env.REACT_APP_DATABASE_URL,
-  projectId: process.env.REACT_APP_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_ID,
-  measurementId: process.env.REACT_MEASUREMENT_ID
-};
-
 export default class Firebase {
-  constructor() {
-    app.initializeApp(config);
-    this.auth = app.auth();
-    this.db = app.firestore();
+  constructor({ db, auth, serverTimestamp }) {
+    this.db = db;
+    this.auth = auth;
+    this.serverTimestamp = serverTimestamp;
   }
 
   setDisplayName = displayName =>
@@ -132,7 +117,7 @@ export default class Firebase {
       .collection("messages")
       .add({
         ...fields,
-        created: app.firestore.FieldValue.serverTimestamp()
+        created: this.serverTimestamp()
       });
 
   getProjectMessages = id =>
