+++ src/index.js
@@ -1,17 +1,37 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import Firebase, { FirebaseContext } from './components/Firebase';
-import * as serviceWorker from './serviceWorker';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import App from "./App";
+import Firebase, { FirebaseContext } from "./components/Firebase";
+import * as serviceWorker from "./serviceWorker";
 
-const firebase = new Firebase()
+import provider from "firebase/app";
+import "firebase/auth";
+import "firebase/firestore";
+
+const config = {
+  apiKey: process.env.REACT_APP_API_KEY,
+  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+  databaseURL: process.env.REACT_APP_DATABASE_URL,
+  projectId: process.env.REACT_APP_PROJECT_ID,
+  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_ID,
+  measurementId: process.env.REACT_MEASUREMENT_ID
+};
+
+provider.initializeApp(config);
+const firebase = new Firebase({
+  db: provider.firestore(),
+  auth: provider.auth(),
+  serverTimestamp: provider.firestore.FieldValue.serverTimestamp
+});
 
 ReactDOM.render(
   <FirebaseContext.Provider value={firebase}>
     <App />
   </FirebaseContext.Provider>,
-  document.getElementById('root'),
+  document.getElementById("root")
 );
 
 // If you want your app to work offline and load faster, you can change
